adjusted Service classes, creating new ExpenseReportService, moving methods from ExpenseService, creating specific Test class and adjusting dependency
