+++ expense-service/src/main/java/com/example/expense/controller/ExpenseReportController.java
@@ -2,7 +2,7 @@ package com.example.expense.controller;
 
 import com.example.expense.dto.ExpenseResponse;
 import com.example.expense.dto.ExpenseSummaryResponse;
-import com.example.expense.service.ExpenseService;
+import com.example.expense.service.ExpenseReportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,13 +21,13 @@ import java.util.List;
 public class ExpenseReportController {
 
     @Autowired
-    private ExpenseService expenseService;
+    private ExpenseReportService expenseReportService;
 
     @GetMapping("/summary/monthly")
     public ResponseEntity<ExpenseSummaryResponse> getMonthlyExpenseSummaryResponse(@RequestParam int year, @RequestParam int month){
         ExpenseSummaryResponse expenseSummaryResponse = new ExpenseSummaryResponse();
 
-        expenseSummaryResponse = expenseService.findExpenseSummaryByMonth(year, month);
+        expenseSummaryResponse = expenseReportService.findExpenseSummaryByMonth(year, month);
 
         return ResponseEntity.status(HttpStatus.OK).body(expenseSummaryResponse);
     }
@@ -36,7 +36,7 @@ public class ExpenseReportController {
     public ResponseEntity<ExpenseSummaryResponse> getAnnualExpenseSummaryResponse(@RequestParam int year){
         ExpenseSummaryResponse expenseSummaryResponse = new ExpenseSummaryResponse();
 
-        expenseSummaryResponse = expenseService.findExpenseSummaryByYear(year);
+        expenseSummaryResponse = expenseReportService.findExpenseSummaryByYear(year);
 
         return ResponseEntity.status(HttpStatus.OK).body(expenseSummaryResponse);
     }
@@ -46,7 +46,7 @@ public class ExpenseReportController {
     public ResponseEntity<List<ExpenseResponse>> getMonthlyExpenses(@RequestParam int year, @RequestParam int month){
         List<ExpenseResponse> expensesFromRequestedMonth = new ArrayList<>();
 
-        expensesFromRequestedMonth = expenseService.findExpensesByMonth(year, month);
+        expensesFromRequestedMonth = expenseReportService.findExpensesByMonth(year, month);
         return ResponseEntity.status(HttpStatus.OK).body(expensesFromRequestedMonth);
 
     }
@@ -54,7 +54,7 @@ public class ExpenseReportController {
     @GetMapping("/detailed/annual")
     public ResponseEntity<List<ExpenseResponse>> getAnnualExpenses(@RequestParam int year){
         List<ExpenseResponse> expensesFromRequestedYear = new ArrayList<>();
-        expensesFromRequestedYear = expenseService.findExpensesByYear(year);
+        expensesFromRequestedYear = expenseReportService.findExpensesByYear(year);
         return ResponseEntity.status(HttpStatus.OK).body(expensesFromRequestedYear);
     }
 }
