+++ expense-service/src/main/java/com/example/expense/service/ExpenseService.java
@@ -115,118 +115,11 @@ public class ExpenseService {
         }
     }
 
-    public ExpenseSummaryResponse findExpenseSummaryByMonth(int year, int month){
-        List<ExpenseResponse> expensesFromGivenMonth = findExpensesByMonth(year,month);
-
-        BigDecimal totalSpentInMonth = findTotalSpentFromGivenPeriod(expensesFromGivenMonth);
-
-        BigDecimal averageDailyExpenses = findAverageDailyExpensesFromMonth(year, month, totalSpentInMonth);
-
-        Map<String, BigDecimal> totalPerCategory = findTotalPerCategory(expensesFromGivenMonth);
-
-        String highestSpentCategory= findHighestSpentCategory(totalPerCategory);
-
-        ExpenseSummaryResponse summaryResponse = new ExpenseSummaryResponse(
-                totalSpentInMonth,
-                averageDailyExpenses,
-                highestSpentCategory,
-                totalPerCategory,
-                expensesFromGivenMonth);
-
-        return summaryResponse;
-    }
-
-    public ExpenseSummaryResponse findExpenseSummaryByYear(int year){
-        List<ExpenseResponse> expensesFromGivenYear = findExpensesByYear(year);
-
-        BigDecimal totalSpentInYear = findTotalSpentFromGivenPeriod(expensesFromGivenYear);
-
-        BigDecimal averageDailyExpenses = findAverageDailyExpensesFromYear(year, totalSpentInYear);
-
-        Map<String, BigDecimal> totalPerCategory = findTotalPerCategory(expensesFromGivenYear);
-
-        String highestSpentCategory= findHighestSpentCategory(totalPerCategory);
-
-        ExpenseSummaryResponse summaryResponse = new ExpenseSummaryResponse(
-                totalSpentInYear,
-                averageDailyExpenses,
-                highestSpentCategory,
-                totalPerCategory,
-                expensesFromGivenYear);
-        return summaryResponse;
-    }
-
-    private BigDecimal findTotalSpentFromGivenPeriod(List<ExpenseResponse> expensesFromGivenMonth) {
-        BigDecimal totalExpensesFromMonth= expensesFromGivenMonth
-                .stream()
-                .map(expenseResponse -> expenseResponse.getValueSpent())
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-        return totalExpensesFromMonth;
-    }
-
-    private BigDecimal findAverageDailyExpensesFromMonth(int year, int month, BigDecimal totalExpensesFromMonth) {
-        BigDecimal averageDailyExpense= totalExpensesFromMonth.divide(BigDecimal.valueOf(getDaysInMonth(year, month)), 2, RoundingMode.HALF_UP);
-        return  averageDailyExpense;
-    }
-
-    private BigDecimal findAverageDailyExpensesFromYear(int year, BigDecimal totalExpensesFromMonth) {
-        BigDecimal averageDailyExpense= totalExpensesFromMonth.divide(BigDecimal.valueOf(getDaysInYear(year)), 2, RoundingMode.HALF_UP);
-        return  averageDailyExpense;
-    }
-
-    private Map<String, BigDecimal> findTotalPerCategory(List<ExpenseResponse> expensesFromGivenMonth) {
-        Map<String, BigDecimal> totalPerCategory = expensesFromGivenMonth.stream()
-                .collect(Collectors.groupingBy(
-                        ExpenseResponse::getCategory,
-                        Collectors.mapping(ExpenseResponse::getValueSpent,
-                                Collectors.reducing(BigDecimal.ZERO, BigDecimal::add))
-                ));
-        return totalPerCategory;
-    }
-
-    private String findHighestSpentCategory(Map<String, BigDecimal> totalPerCategory) {
-        String highestSpentCategory = totalPerCategory.entrySet().stream()
-                .max(Map.Entry.comparingByValue())
-                .map(Map.Entry::getKey)
-                .orElse("N/A");
-        return highestSpentCategory;
-    }
 
 
-    public List<ExpenseResponse> findExpensesByMonth(int year, int month){
-        LocalDate startDate = LocalDate.of(year, month,1);
-        LocalDate endDate = LocalDate.of(year, month, getDaysInMonth(year, month));
-        try {
-            return findVerifiedExpensesBetweenDates(startDate, endDate);
-        }
-        catch (Exception e){
-            throw new ExpenseNotFoundException(String.format("Failed to find expenses from Year %d, Month %d", year, month));
-        }
-    }
 
-    public List<ExpenseResponse> findExpensesByYear(int year){
-        LocalDate startDate = LocalDate.of(year, 01,1);
-        LocalDate endDate = LocalDate.of(year, 12, getDaysInMonth(year, 12));
-        try {
-            return findVerifiedExpensesBetweenDates(startDate, endDate);
-        }
-        catch (Exception e){
-            throw new ExpenseNotFoundException(String.format("Failed to find expenses from Year %d", year));
 
-        }
-    }
 
-    private List<ExpenseResponse> findVerifiedExpensesBetweenDates(LocalDate startDate, LocalDate endDate) {
-        return expenseRepository.findByDateBetween(startDate, endDate).stream().map(expense -> mapExpenseToExpenseResponse(expense)).toList();
-    }
-
-    public int getDaysInMonth(int year, int month) {
-        return YearMonth.of(year, month).lengthOfMonth();
-    }
-
-    public int getDaysInYear(int year) {
-    return Year.of(year).length();
-}
 
     public void removeExpense(Long id){
         if(!expenseRepository.existsById(id)){
